*  Add a timestamp policy to the glom parameters item. *  Make ringselector use the item header size to figure out how big    the item is. *  Ensure that all unit tests pass. *  Get rid of some of the TODO's that we won't DO.

+++ sbs/readout/CExperiment.cpp
@@ -568,7 +568,7 @@ CExperiment::DocumentPackets()
     CRingTextItem item(NULL_TIMESTAMP, m_nSourceId, BARRIER_NOTBARRIER,
                        PACKET_TYPES, packetDefs,
 		       offset,
-		       now);
+		       static_cast<uint32_t>(now));
     item.commitToRing(*m_pRing);
   }
 }
